pyproject.toml
[project]
name = "mail-rememberer"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "mistralai>=1.8.1",
    "postmarker>=1.0",
]

[dependency-groups]
dev = [
    "python-dotenv>=1.1.0",
]
